+++ src/bops-functions/prebuilt-functions/flux-control/if.ts
@@ -24,8 +24,8 @@ export const ifBopsFunctionInformation : InternalMetaFunction = {
   description: "Returns one of two given values based on whether the given boolean is true or false",
   inputParameters: {
     boolean : { type: "boolean", required: true },
-    valueIfTrue : { type: "any", required: true },
-    valueIfFalse : { type: "any", required: false },
+    ifTrue : { type: "any", required: true },
+    ifFalse : { type: "any", required: false },
   },
   outputData: {
     outputValue: { type: "any", required: true },