+++ src/configuration/protocols/protocol-config-type.ts
@@ -1,7 +1,0 @@
-import { CloudedObject } from "common/types/clouded-object";
-
-export interface ProtocolConfigType {
-  protocolType : string;
-  protocolVersion : string; // SemVer;
-  configuration : CloudedObject;
-}