+++ test/bops-functions/bops-engine/test-data/test-system.ts
@@ -4,6 +4,7 @@ import { internalBop } from "./business-operations/internal-bop";
 import { schemaBop } from "./business-operations/schema-bop";
 import { externalBop } from "./business-operations/external-bop";
 import { variableBop } from "./business-operations/variables-bop";
+import { packageBop } from "./business-operations/package-bop";
 
 export const testSystem : ConfigurationType = {
   name: "test-system",
@@ -25,16 +26,9 @@ export const testSystem : ConfigurationType = {
     schemaBop,
     externalBop,
     variableBop,
+    packageBop,
   ],
   protocols: [
-    {
-      protocolVersion: "",
-      protocolType: "HTTP_JSONBODY",
-      configuration: {
-        port: 8080,
-        routes: [],
-      },
-    },
   ],
 };